+++ examples/update-exchange-rate.rs
@@ -1,15 +1,14 @@
-#![allow(deprecated)]
 //! An example showing how to do a simple update of the exchange rate between
 //! CCD and EUR.
 use anyhow::Context;
 use clap::AppSettings;
 use concordium_rust_sdk::{
     common::types::TransactionTime,
-    endpoints::{self, Endpoint},
     types::{
         transactions::{update, BlockItem, Payload},
         BlockSummary, ExchangeRate, TransactionStatus, UpdateKeyPair, UpdatePayload,
     },
+    v2::{self, BlockIdentifier, ChainParameters},
 };
 use std::path::PathBuf;
 use structopt::StructOpt;
@@ -19,15 +18,9 @@ struct App {
     #[structopt(
         long = "node",
         help = "GRPC interface of the node.",
-        default_value = "http://localhost:10000"
+        default_value = "http://localhost:20000"
     )]
-    endpoint: Endpoint,
-    #[structopt(
-        long = "rpc-token",
-        help = "GRPC interface access token for accessing all the nodes.",
-        default_value = "rpcadmin"
-    )]
-    token:    String,
+    endpoint: v2::Endpoint,
     #[structopt(long = "key", help = "Path to update keys to use.")]
     keys:     Vec<PathBuf>,
 }
@@ -49,19 +42,15 @@ async fn main() -> anyhow::Result<()> {
         })
         .collect::<anyhow::Result<_>>()?;
 
-    let mut client = endpoints::Client::connect(app.endpoint, app.token).await?;
-
-    let consensus_status = client
-        .get_consensus_status()
-        .await
-        .context("Could not obtain status of consensus.")?;
+    let mut client = v2::Client::new(app.endpoint).await?;
 
     // Get the key indices, as well as the next sequence number from the last
     // finalized block.
-    let summary: BlockSummary = client
-        .get_block_summary(&consensus_status.last_finalized_block)
+    let summary: ChainParameters = client
+        .get_block_chain_parameters(BlockIdentifier::LastFinal)
         .await
-        .context("Could not obtain last finalized block")?;
+        .context("Could not obtain last finalized block's chain parameters")?
+        .response;
 
     // find the key indices to sign with
     let signer = summary
@@ -69,20 +58,11 @@ async fn main() -> anyhow::Result<()> {
         .construct_update_signer(&summary.common_update_keys().micro_gtu_per_euro, kps)
         .context("Invalid keys supplied.")?;
 
-    let seq_number = match &summary {
-        BlockSummary::V0 { data, .. } => {
-            data.updates
-                .update_queues
-                .micro_gtu_per_euro
-                .next_sequence_number
-        }
-        BlockSummary::V1 { data, .. } => {
-            data.updates
-                .update_queues
-                .micro_gtu_per_euro
-                .next_sequence_number
-        }
-    };
+    let seq_number = client
+        .get_next_update_sequence_numbers(BlockIdentifier::LastFinal)
+        .await?
+        .response;
+    let seq_number = seq_number.micro_ccd_per_euro;
 
     let effective_time = 0.into(); // immediate effect
     let timeout =
@@ -103,7 +83,7 @@ async fn main() -> anyhow::Result<()> {
     loop {
         interval.tick().await;
         match client
-            .get_transaction_status(&submission_id)
+            .get_block_item_status(&submission_id)
             .await
             .context("Could not query submission status.")?
         {