+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## Unreleased changes
 
 - The sdk now requires a `rustc` version at least 1.67 (Before it required version 1.66).
-- Add Display implementation to BlockIdentifier.
+- Add `Display` implementation to `BlockIdentifier`.
+- Add `Display` and `FromStr` implementations for `AccountIdentifier`.
+
 
 ## 3.1.0